tests: split smoke.spec.ts by theme + extract shared helpers
The single 2447-line monolith grew organically over many PRs. Split
into 9 thematic spec files matched to the natural test clusters,
with shared launch/mockMainIpc/activePane/waitForLogEntry/callTestIpc
helpers in tests/_helpers.ts (underscore prefix keeps playwright
from treating it as a test file).

  tests/_helpers.ts             — launch + mocks + IPC test bridge
  tests/preload.spec.ts         — preload API surface
  tests/create-flow.spec.ts     — New workspace modal, validation, OAuth
  tests/mock-mode.spec.ts       — MOCK MODE + selection + oauth cmd
  tests/workspace-lifecycle.spec.ts — pause/resume/close/start/past, hamburger
  tests/multi-session.spec.ts   — tab strip + sessions.json roundtrip
  tests/always-mount.spec.ts    — pty:attach + isolation + CSS cascade
  tests/jsonl-watcher.spec.ts   — real DB + chokidar ingest behavior
  tests/broker-sessions.spec.ts — broker→claude mapping pipeline
  tests/observability.spec.ts   — pane content, slot consumers, live push
  tests/image-library.spec.ts   — image picker filter

Same 44 tests, same coverage. Largest file dropped from 2447 to 540
lines (observability — slot consumers + ObservabilityPane + live push
keep a lot of mock fixture data in line). All 44 e2e + 40 unit pass.
