+++ tests/preload.spec.ts
@@ -1,0 +1,28 @@
+import { test, expect } from '@playwright/test';
+import { launch } from './_helpers.js';
+
+test('preload exposes window.api with all expected surfaces', async () => {
+  const { app, window } = await launch();
+  try {
+    const types = await window.evaluate(() => ({
+      api: typeof (window as unknown as { api?: unknown }).api,
+      workspace: typeof (window as unknown as { api?: { workspace?: unknown } }).api?.workspace,
+      images: typeof (window as unknown as { api?: { images?: unknown } }).api?.images,
+      vault: typeof (window as unknown as { api?: { vault?: unknown } }).api?.vault,
+      pty: typeof (window as unknown as { api?: { pty?: unknown } }).api?.pty,
+      fs: typeof (window as unknown as { api?: { fs?: unknown } }).api?.fs,
+      dialog: typeof (window as unknown as { api?: { dialog?: unknown } }).api?.dialog
+    }));
+    expect(types).toEqual({
+      api: 'object',
+      workspace: 'object',
+      images: 'object',
+      vault: 'object',
+      pty: 'object',
+      fs: 'object',
+      dialog: 'object'
+    });
+  } finally {
+    await app.close();
+  }
+});
